% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcproj-package.R
\docType{package}
\name{cmcproj-package}
\alias{cmcproj}
\alias{cmcproj-package}
\title{cmcproj: Statistical Research Project Template}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an opinionated project template for applied statistical research. Designed to include both paper, replication code, and standalone software package.
}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})

}
\keyword{internal}
